requirements.txt
streamlit
insightface==0.7.3
onnxruntime==1.18.0
numpy==1.26.4
opencv-python-headless
Pillow
